% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm.R
\name{AIC_quasi}
\alias{AIC_quasi}
\title{Calculates AIC from a quasi-binomial}
\usage{
AIC_quasi(model)
}
\arguments{
\item{model}{A generalized linear model of a quasibinomial}
}
\value{
An invisible NULL
}
\description{
This function inputs a quasi binomial glm model and outputs a quasi-AIC score
. All credits for this code goes to Mario Cortina Borja.
}
