% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotype_imputated_df.R
\name{check_strand_concordance}
\alias{check_strand_concordance}
\title{Check strand concordance}
\usage{
check_strand_concordance(set1, set2)
}
\arguments{
\item{set1}{with two alleles}

\item{set2}{with another two alleles}
}
\value{
boolean, TRUE or FALSE depending on whether
}
\description{
Check strand concordance
}
