% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_analysis.R
\name{Coeff.OR2}
\alias{Coeff.OR2}
\title{Calculates OR's, SF's and CI's from a glm.}
\usage{
Coeff.OR2(mod1, n.digits = 5, verbose = F, rounded = F)
}
\arguments{
\item{mod1}{the glm model}

\item{n.digits}{the number of significant digits to output}

\item{verbose}{whether it should print an output}

\item{rounded}{should the pvalues be rounded or not}
}
\value{
the summary.glm(model) updated with the CI and the OR/SF.
}
\description{
This function, given a logistic regression model modifies the normal output
of summary(model) to return also a 95\% CI (upper and lower bounds) and the
OR/SF of the association. All credits for this code goes to Mario Cortina
Borja.
}
