% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complement_alleles.R
\name{complement_alleles}
\alias{complement_alleles}
\title{Get the complementary allele sequence}
\usage{
complement_alleles(sequence, mode = "dna")
}
\arguments{
\item{sequence}{character vector, with a nucleotide sequence.}

\item{mode}{string. If "dna" it complements "A" to "T". If "rna" it
complements "A" to "U".}
}
\value{
character vector
}
\description{
For a character vector of a genomic sequences it returns a character vector
with the complementary sequence.
}
