% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_paths}
\alias{create_paths}
\title{Assembles a path and creates intermediary directories}
\usage{
create_paths(parts)
}
\arguments{
\item{parts}{character vector}
}
\value{
string
}
\description{
Given a character vector, it creates a non-OS specific path and makes all
intermediate directories necessary.
}
\examples{
create_paths(c("All", "the", "way", "to", "my", "directory"))
}
