% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSummary.R
\name{createSummary}
\alias{createSummary}
\title{Create Summary}
\usage{
createSummary(variable, group)
}
\arguments{
\item{variable}{a variable}

\item{group}{a group}
}
\value{
a summary table
}
\description{
Create Summary
}
