% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EP-package.R
\docType{package}
\name{EP-package}
\alias{EP}
\alias{EP-package}
\title{EP: Analysis of epistatic interactions for the Epistasis Project dataset}
\description{
R pipeline that can be used to study epistatic interactions in 
  present and future genetic datasets of case-control cohorts.
}
\author{
\strong{Maintainer}: Armand González \email{armandgonzalezescalante@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
