% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_inheritances.R
\name{generate_inheritances}
\alias{generate_inheritances}
\title{Generate inheritances for a given set of snps in a dataset.}
\usage{
generate_inheritances(snps, data)
}
\arguments{
\item{snps}{SNP_set object}

\item{data}{dataframe, where \code{\link[base]{names}(snps)} corresponds to
variables in data.}
}
\value{
The same dataframe as in data but with 3 * \code{length(snps)} new columns added to it. Once for each inheritance model per snp.
}
\description{
Given a SNP_set object and a dataframe with the genotypes for each case for
those snps, recodify in new variables the genotypes into each one of three
inheritance models (additive, dominant and recessive.)
}
