% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{generate_list_of_snp_pairs}
\alias{generate_list_of_snp_pairs}
\title{Generates a list of snp pairs to be used in further analysis steps}
\usage{
generate_list_of_snp_pairs(mode, snp_df)
}
\arguments{
\item{mode}{string, whether it should use a snp model in the interactions
list ("reported_model") or just the snp in question ("best_model")}

\item{snp_df}{dataframe, the snp with the snp interactions with the variable
"snp1" being the first member of the interaction and "snp2" the second. If
using models, "model1" should define the first member of the interaction and
model2 the second one.}
}
\value{
A list of character vectors, containing the character vectors the snp
names or snp models interactions, depending on the mode of operation selected
}
\description{
Generates a list of snp pairs with (reported_model) or without (best_model)
the snp model.
}
