% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_object.R
\name{generate_object}
\alias{generate_object}
\title{Creates a list of SNP objects}
\usage{
generate_object(exist, snps, ...)
}
\arguments{
\item{exist}{logical, does the list of SNP objects already exist?}

\item{snps}{character vector with "gene_snp" names in the case of genotyped
df (used to retrieve the gene names) or just RefSNP id's if imputated.}

\item{...}{further arguments passed to or from other methods.
Use argument dataset, for inputting a dataframe, with all the SNP data and
argument verbose, logical, for whether the function should print output.}
}
\value{
a SNP_set S3 object (a list of SNP objects)
}
\description{
Given a data frame, it checks if a list of SNP objects exist and if not
creates it with a set of given snp id's. If it does exist it expects that
the data frame inputted is an imputated one and for each SNP adds its
attributes to the SNP object.
}
