% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputated.R
\name{genotype_imputated_df}
\alias{genotype_imputated_df}
\title{Recodifies an imputated dataframe into genotypes}
\usage{
genotype_imputated_df(list_of_objects, df, thousand_genomes, verbose, ...)
}
\arguments{
\item{list_of_objects}{'SNP_set' S3 class, contains several SNP objects with
their respective major and minor allele that will be used to see if they
match the reference and alternative allele in the reference panel
(thousand genomes)}

\item{df}{dataframe, with the imputated data for each snp under study}

\item{thousand_genomes}{data.frame, which contains the reference panel
reference and alternative allele with respect to which the imputation took
place}

\item{verbose}{logical, should the function print to output the steps it
performs?}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A dataframe with the imputation converted into genotypes.
}
\description{
For a given imputated dataframe, it recodifies the output obtained into
genotype data according to a reference panel and checks that it matches the
same strand in the genotyped dataframe through the list_of_objects.
}
