% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model.R
\name{get_model}
\alias{get_model}
\title{Gets the corresponding inheritance model}
\usage{
get_model(snps, models, master_list)
}
\arguments{
\item{snps}{character vector, of snp id's}

\item{models}{character vector, with the inheritance model of the snps}

\item{master_list}{list, with all results from the analysis}
}
\value{
a vector with all the snp_id + model
}
\description{
This function gets a string with the corresponding inheritance model for each
SNP (dominant, recessive or additive) and returns the id of the SNP with the
first letter of the model (d, r or a) added to it. If the inheritance model
string is "unknown", it looks for the best model at the master_list object
and retrieves it.
}
