% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_packages}
\alias{load_packages}
\title{Load and attach libraries from vector of names.}
\usage{
load_packages(packages)
}
\arguments{
\item{packages}{character vector, with the names of the libraries to be
loaded}
}
\value{
invisible NULL
}
\description{
Convenience function that loads and attaches all libraries named in a
character vector silently.
}
\examples{
packages <- c("ggplot2", "xlsx")
load_packages(packages)
}
