% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_heterozygotes.R
\name{order_heterozygotes}
\alias{order_heterozygotes}
\title{Order heterozygotes by alphabetical order}
\usage{
order_heterozygotes(x)
}
\arguments{
\item{x}{character vector, consisting on the genotypes we want to order
alphabetically}
}
\value{
character vector, ordered list of genotypes
(the heterozygotes are ordered alphabetically)
}
\description{
Order heterozygotes by alphabetical order
}
