% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_analysis.R
\name{perform_analysis}
\alias{perform_analysis}
\title{Main function used to perform glm models to study association in epistasis
related projects}
\usage{
perform_analysis(
  .mode,
  .submode = NA,
  .data,
  snps,
  covariates,
  .verbose = verbose
)
}
\arguments{
\item{.mode}{string, whether we want to study the "main_effects",
the "interaction" with covariates or the "snp" - snp interactions.}

\item{.submode}{if the mode chosen is the "snp" one you can further declare
wether you want to study the "best_model" for the best performing accoring to
the AIC score obtained by main_effects or just the "reported_model" if you
are inputting the model too.}

\item{.data}{character vector, names of the datasets with all the
information of the inheritance models.}

\item{snps}{SNP_set object with all the SNPs information}

\item{covariates}{the covariates to control for while performing the glm.}

\item{.verbose}{whether it should print information to the output while
running.}
}
\value{
a nested list with all the information on the analyses performed
results
}
\description{
Main function used to perform glm models to study association in epistasis
related projects
}
