% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_object.R
\name{print.SNP}
\alias{print.SNP}
\title{Default print for SNP_object}
\usage{
\method{print}{SNP}(SNP)
}
\arguments{
\item{SNP}{SNP object}
}
\value{
invisible NULL
}
\description{
Method for the SNP object that is dispatched when the print generic is used on
a SNP object
}
