% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_object.R
\name{print.SNP_set}
\alias{print.SNP_set}
\title{Default print for SNP_set object}
\usage{
\method{print}{SNP_set}(SNP_set_object)
}
\arguments{
\item{SNP_set_object}{SNP_set object}
}
\value{
invisible NULL
}
\description{
Method for the SNP_set object that is dispatched when the print generic is used on
a SNP_set object
}
