% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_results.R
\name{print_INT}
\alias{print_INT}
\title{Print interaction between SNP and covariates Age and Sex.}
\usage{
print_INT(list, covariates, corrected = F)
}
\arguments{
\item{list}{list, containing the results in the analysis (i.e. master_list)}

\item{covariates}{vector, with the covariates that we want to study}

\item{corrected}{boolean, whether it should correct for multiple testing.
Default is false.}
}
\value{
NULL (invisible)
}
\description{
This function prints covariate interaction results, applying for multiple
testing correction or not. In particular prints the snp-covariate interaction
with its p-value and the genes to which they correspond.
}
