% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_results.R
\name{print_ME}
\alias{print_ME}
\title{Prints main effects results}
\usage{
print_ME(list, corrected = F)
}
\arguments{
\item{list}{list, containing the results in the analysis (i.e. master_list)}

\item{corrected}{boolean, whether it should correct for multiple testing.
Default is false.}
}
\value{
NUll (invisible)
}
\description{
This function prints main effects results, applying multiple testing
correction or not. In particular prints the snps with their p-values and the
genes to which they correspond.
}
