% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_results.R
\name{print_SNP_SNP}
\alias{print_SNP_SNP}
\title{Print SNP-SNP interactions results.}
\usage{
print_SNP_SNP(master_list, list_of_snp_pairs, list_of_APOE, corrected = F)
}
\arguments{
\item{master_list}{list, containing the results in the analysis}

\item{list_of_snp_pairs}{list, containing the SNP-SNP interactions
to look for}

\item{list_of_APOE}{list, containing the SNP-APOE4 interactions to look for}

\item{corrected}{boolean, whether it should correct for multiple testing.
Default is false.}
}
\value{
NULL (invisible)
}
\description{
This function prints SNP-SNP interaction results, applying for multiple
testing correction or not. In particular prints the snp-snp interaction
with its p-value and the genes to which they correspond.
}
