% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rec_list}
\alias{rec_list}
\title{Create empty recursive list from a numerical vector}
\usage{
rec_list(x)
}
\arguments{
\item{x}{numerical vector}
}
\value{
list
}
\description{
This function creates and empty list for a numerical vector so that every
position corresponds to the number of elements inside that specific recursion
level being \code{length(vector)} the number of recursions levels of the list.
}
\examples{
rec_list(c(2,3))
}
