% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{run_shell}
\alias{run_shell}
\title{Run programs in the shell}
\usage{
run_shell(command)
}
\arguments{
\item{command}{character string, quoted command}
}
\value{
invisible NULL
}
\description{
Wrapper around the system() function that works on Rmarkdown
}
