% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_interactions.R
\name{store_interactions}
\alias{store_interactions}
\title{Store interactions of SNP-SNP}
\usage{
store_interactions(snp_dataset, master_list)
}
\arguments{
\item{snp_dataset}{snp dataframe}

\item{master_list}{master list with all results}
}
\value{
data.frame, the master_list with all updated results
}
\description{
Store interactions of SNP-SNP
}
