% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_object.R
\name{str.SNP_set}
\alias{str.SNP_set}
\title{Default str for SNP_set object}
\usage{
\method{str}{SNP_set}(SNP_set_object)
}
\arguments{
\item{SNP_set_object}{SNP_set object}
}
\value{
invisible NULL
}
\description{
Method for the SNP_set object that is dispatched when the str generic is used on
a SNP_set object
}
