% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsplit2vector.R
\name{strsplit2vector}
\alias{strsplit2vector}
\title{Split string into a vector of components}
\usage{
strsplit2vector(string, pattern)
}
\arguments{
\item{string}{string}

\item{pattern}{string}
}
\value{
vector
}
\description{
Wrapper around \code{\link[base]{strsplit}}, returning a vector instead of a list.
}
